Add workflow transformer to be used in use-chat
