+++ .changeset/wild-lemons-buy.md
@@ -1,0 +1,5 @@
+---
+'@mastra/ai-sdk': patch
+---
+
+Add transformer to support workflows in useChat
